+++ FA-Backend-Project/src/main/java/com/example/febackendproject/DTO/PartialProductDTO.java
@@ -18,7 +18,7 @@ public class PartialProductDTO {
     
     private String saleUnit;
     private Double saleUnitPrice;
-    private Integer measurePerSaleUnit;
+    private Double measurePerSaleUnit;
     
     private Integer discountPercentage;
     private Double discountedPrice;
