+++ FA-Backend-Project/src/main/java/com/example/febackendproject/DTO/PartialProductStockDTO.java
@@ -19,6 +19,6 @@ public class PartialProductStockDTO {
     private Double measurePrice;
     
     private String saleUnit;
-    private Integer measurePerSaleUnit;
+    private Double measurePerSaleUnit;
 
 }
